Fix compilation errors from transfer

+++ include/key_aware/key_aware.h
@@ -14,11 +14,11 @@ public:
 
     StringView (const char * str) noexcept :
         m_ptr(str),
-        m_length(strlen(str))
+        m_length((int32_t)strlen(str))
     {
     }
 
-    StringView (const char * str, const size_t length) noexcept :
+    StringView (const char * str, const int32_t length) noexcept :
         m_ptr(str),
         m_length(length)
     {
@@ -27,7 +27,7 @@ public:
     StringView (const StringView &) noexcept = default;
     StringView & operator= (const StringView &) noexcept = default;
 
-    size_t Length () const noexcept {
+    int32_t Length () const noexcept {
         return m_length;
     }
 
@@ -35,13 +35,13 @@ public:
         return m_ptr;
     }
 
-    const char & operator[] (const size_t index) const noexcept {
+    const char & operator[] (const int32_t index) const noexcept {
         return m_ptr[index];
     }
 
 private:
     const char * m_ptr;
-    size_t       m_length;
+    int32_t      m_length;
 };
 
 int32_t Distance (const StringView & from, const StringView & to);
